Always update cached file on dav_lookup
on open/read/close sequence, fuse_lookup is called before fuse_read to get file
attributes, it might be possible that the cached file contains older size attribute, and
so the call to fuse_read doesn't read the full content of the file.

So for that, on dav_lookup we update the cached file attributes (bug #61049).