Optimize server side file size change detection
Commit 43ef5ac7b2f2 introduces performance issue, as the
update_cache_file is downloading the whole file on each lookup. It is
enough to get the latest file size when a file is not open or not dirty
to detect if it has been changed on the server, so next FUSE read will
have the correct size.

(Fixes 43ef5ac7b2f2 Always update cached file on dav_lookup)